mlx4: fix error case in mlx4_tx_burst()
Substracting the WR index by the number of packets to be sent makes no
sense. Instead of the WR index, the number of sent packets must be
determined when ibv_post_send() fails.

Moreover, the number of completions events can now be 0. In that case,
some WRs can still be in use. Fix related assert() conditions.

Acked-by: Damien Millescamps <damien.millescamps@6wind.com>
Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>