mlx4: reduce cache misses in TX path
Part of performance improvements. This is TX path. The goal is to avoid
cache misses. The txq_elt was changed to two structures txq_wr and txq_buf.

As for RX flow, we dont need to keep wr after post_send, therefore we can
use only burst_size elements of tx_elt_wr array. The impact of this patch on
performance is very big.

AM:
- Coding style changes.
- Also in txq_complete(), make sure all completed mbuf pointers are set to
  NULL, because wr->num_sge isn't easily available anymore.
- In mlx4_tx_burst(), bring back rollback to replace abort() in case of
  ibv_post_send() failure.

Signed-off-by: Olga Shern <olgas@mellanox.com>
Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>