mlx4: fix and improve MTU/MRU handling
The problem when configuring MTU through mlx4_dev_set_mtu() is basically
that it also sets the maximum RX size to the interface MTU, which doesn't
account for Ethernet headers. This makes the PMD unable to receive MTU-sized
packets.

This commit also changes the way the maximum RX size is derived from the MTU
to toggle scattered packets support more efficiently.

Acked-by: Guillaume Gaudonville <guillaume.gaudonville@6wind.com>
Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>