mlx4: do not reallocate mbufs nor queues when updating MTU
Depending on its size (larger or lower than a single mbuf), the new MTU
makes the PMD automatically enable or disable scattered RX mode.
This is currently done by destroying and recreating all RX queues.

There are two issues with this approach:
- Due to how RSS works internally, there is no guarantee that the new queues
  will be part of the RSS pool and may not receive packets.
- mlx4_dev_set_mtu() may be called from a control thread which does not have
  a mempool cache. Freeing mbufs in these conditions is not supported and
  may cause data corruption and crashes.

To prevent this, this commit adds rxq_rehash() which recycles mbufs and is
able to reset individual queues without destroying them. It is also lighter
and less likely to fail.

Acked-by: Guillaume Gaudonville <guillaume.gaudonville@6wind.com>
Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>