mlx4: prevent mlx4_rx_burst() from returning duplicate mbufs
To use less memory, the last RX path optimization removed the mbuf pointer
from struct rxq_elt by storing it as the work request ID (wr_id field from
struct ibv_recv_wr).

Each packet received by mlx4_rx_burst() causes the associated mbuf to be
removed and replaced with a fresh one. This worked correctly before the
optimization because the WR ID was the index in the array of the RX queue,
but since then, this information is not available and duplicate buffers are
often returned by mlx4_rx_burst(), causing packet corruption and crashes.

The reason is mlx4_rx_burst() now assumes that each received burst starts
with the first entry of the RX queue, thus uses the same array index for
work completions and elements in the RX queue. This is wrong, as not only
the adapter may receive a second burst while the first one is being
processed (the first index of this burst won't be 0), but also buffers are
filled by the hardware in the order they are submitted, which means that
RX queue indices above the size of the current burst will be used for the
next burst anyway.

This commit solves this issue while keeping the size optimization of
struct rxq_elt by combining the RX queue index and the negative mbuf data
offset as the work request ID. The mbuf pointer can then be derived from the
address of the submitted buffer.

Acked-by: Guillaume Gaudonville <guillaume.gaudonville@6wind.com>
Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>