mlx4: add software support for any number of TX segments
The maximum number of segments per packet allowed for TX is limited to
MLX4_PMD_SGE_WR_N, currently defined as 4. This is not enough for many use
cases.

Unfortunately, increasing it past this value and asking the hardware for
more causes a considerable performance drop that also affects packets with
few segments.

As a workaround, this commit implements software linearization of extra
segments.

Acked-by: David Marchand <david.marchand@6wind.com>
Signed-off-by: Adrien Mazarguil <adrien.mazarguil@6wind.com>